+++ ChangeLog.md
@@ -1,6 +1,11 @@
 ## Change log
+
 - Version: v0.5.30
   - Remove donate bitcoin
+  - Add build script
+  - Release
+    - Add arm64
+    - App AppImage
 
 - Version: v0.5.29
   - Use [RabbitCommon v2.3.2](https://github.com/KangLin/RabbitCommon/releases/tag/v2.3.2)